WIP: MVP-like model and TUI final implementation
After many days of testing we may have figured out what should be
the pattern to build the Terminal UI. A modified version of an MVP
pattern is used in a simplified manner, since the view is extremely
dumb in our code. It only grabs data from the user, there's minimal
iteration with the user after the data is filled.