[CLOYSTER-131] Fix ISO download and the progress bar widget (#84)
* service: throw an exception when we cannot find a daemon

We create a exception exclusive for daemon related errors

* Begin fixing ISO download on node OS install step

Now, we effectively download the ISOs. We ask for some recent versions
also

* functions: Allow the command proxy to get from stdout or stderr

wget prints its status in stderr for some reason, we need to account
for that.

We also need to get "line" separators different from \n, like \r. They
would be less frequent, but we would need them

* Fix the progress dialog box when downloading an ISO

* Fix some Sonarcloud issues

* fixup! Fix some Sonarcloud issues

* Fix progress bar not showing message and not closing after ending

* fixup! Fix progress bar not showing message and not closing after ending