Make Cluster generic over Runner
This commit adds a type parameter in Cluster class regarding
the Runner. This is in an intetion to decouple cluster
from the Runner in the near future. Then all these type
parameters spreding all over the place can be removed

Runner is a special case here, it what runs commands so
I want to be very strict in where we can use it.

Another thing to pay attention is that Cluster is a kind
of "god" object where every other parts of the code rely
into, we want to break this into smaller pieces and have
the code depending on more atomic parts, this will help
with testability.